new character class kits
