+++ src/App.tsx
@@ -28,41 +28,49 @@ import { getSavedCharacterData } from "./utils/getSavedCharacterData";
  * - [x] needs assassin skills icon: bloody sword
  * - [x] barbarian: hunting horn, totem, brutal helm, barbarian, biceps, ancient sword
  * - [x] needs barbarian skills icon: snake totem
- * - [] needs barbarian equipment (maybe give totemic/tribal items?)
+ * - [x] needs barbarian equipment (maybe give totemic/tribal items?)
+ * - [x] barbarian needs long bows, warhammer
  * - [x] bard
  * - [x] needs some random scrolls from druid spell list in equipment
  * - [x] druid: wizard staff
- * - [] needs druid equipment (wooden shields, holy symbol)
- * - [] spell scrolls for druid spells
+ * - [x] needs druid equipment (wooden shields, holy symbol)
+ * - [x] spell scrolls for druid spells
  * - [x] illusionist: pointy hat
- * - [] needs illusionist equipment (just like magic user but include illusionist scrolls)
- * - [] illusionist equipment, some need a quartz (10gp)
+ * - [x] needs illusionist equipment (just like magic user but include illusionist scrolls)
+ * - [x] illusionist equipment, some need a quartz (10gp)
+ * - [x] illusionist and magic user need some silver daggers
  * - [x] knight: visored helm
- * - [] needs knight equipment (no leather armor, no missile weapons)
+ * - [x] needs knight equipment (no leather armor, no missile weapons)
  * - [x] paladin: templar shield
- * - [] needs paladin equipment (holy symbol)
+ * - [x] needs paladin equipment (holy symbol)
  * - [x] ranger: compass, arrow cluster, high shot
- * - [] needs ranger equipment (more long bows)
+ * - [x] needs ranger equipment (more long bows)
  * - [x] needs tracking table icon: paw print
  *
  *  4/6 new demihuman race-classes remain (icons):
  * - [x] duergar: heavy helm, dwarf face
  * - [x] needs same equipment check as dwarves
  * - [x] gnome: wizard face
- * - [] needs illusionist/gnome equipment (include illusionist scrolls)
+ * - [x] needs gnome equipment (include illusionist scrolls)
  * - [x] needs random burrowing mammal pets (resolve like bard instruments)
  * - [x] half-orc: orc head
- * - [] needs half-orc equipment
+ * - [x] needs half-orc equipment
  * - [x] svirfneblin: war pick
  * - [x] needs same equipment check as dwarves
  *
+ * - [] Check alignment -> paladin, druid, and ranger
+ * - [x] War Hammer -> Warhammer
  * - [x] FIGHTER NEEDS weapon specialization rules
  * - [x] capitalize psionicst skills
  * - [] Fix wild magic-user spell level variability
  * - [x] Add level one illusionist spells and some high level scrolls
  * - [x] BARD NEEDS getExperienceAdjustment added for CHA
+ * - [] Update weapons and armor OSE pages to include slots instead of
+ * - weight in coins (which is awkward), Make this the new "purchase sheet",
+ * - combine with other various items in Knave. Club/Cudgel, Flail/Mace/Morning Star, Halberd/Polearm
+ * - [] Add store items once pdf is created so images can be simply inserted into app as a reference
  *
- * - Add custom kobold class, GiFoxHead (like an infravision halfling)
+ * - Add custom kobold class, GiFoxHead (like an infravision halfling)?
  * @see https://dysonlogos.blog/2011/04/13/tuckers-kobolds/
  * - Create own random treasure app tool
  *  - Consider simple random spellbook generator for Illusionist and Magic-User
