+++ src/characterData/classes/elf.ts
@@ -1,6 +1,6 @@
 import React from "react";
 import { GiElfHelmet } from "react-icons/gi";
-import { magicUserEquipment } from "../equipmentKits";
+import { magicUserEquipment } from "../equipmentKits/magicUserEquipment";
 
 export const elf = {
   name: "Elf",
