+++ src/characterData/equipmentKits/assassinEquipment.ts
@@ -1,0 +1,125 @@
+/**
+ * Assassin equipment kits cannot be higher than 9 slots.
+ * Chance of starting with poison.
+ * 9 + starting gold = 10
+ */
+export const assassinEquipment = [
+  [
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Thieves' Tools</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Thieves' Tools</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Thieves' Tools</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>Crossbow</strong> (Damage: 1d6, Qualities: Missile (5’-80’ / 81’-160’ / 161’-240’), Slow, Two-handed, 3 slots)",
+    "<strong>Case With Bolts</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Thieves' Tools</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Crossbow</strong> (Damage: 1d6, Qualities: Missile (5’-80’ / 81’-160’ / 161’-240’), Slow, Two-handed, 3 slots)",
+    "<strong>Case of Bolts</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Thieves' Tools</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Short Bow</strong> (Damage: 1d6, Qualities: Missile (5’-50’ / 51’-100’ / 101’-150’), Two-handed, 2 slots)",
+    "<strong>Quiver of Arrows</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Thieves' Tools</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>Crossbow</strong> (Damage: 1d6, Qualities: Missile (5’-80’ / 81’-160’ / 161’-240’), Slow, Two-handed, 3 slots)",
+    "<strong>Silver Tipped Bolts</strong> (d4 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Short Bow</strong> (Damage: 1d6, Qualities: Missile (5’-50’ / 51’-100’ / 101’-150’), Two-handed, 2 slots)",
+    "<strong>Quiver of Arrows</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Thieves' Tools</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Crossbow</strong> (Damage: 1d6, Qualities: Missile (5’-80’ / 81’-160’ / 161’-240’), Slow, Two-handed, 3 slots)",
+    "<strong>Case of Bolts</strong> (d20 usage die)",
+    "<strong>Silver Tipped Bolts</strong> (d4 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>Short Bow</strong> (Damage: 1d6, Qualities: Missile (5’-50’ / 51’-100’ / 101’-150’), Two-handed, 2 slots)",
+    "<strong>Quiver of Arrows</strong> (d20 usage die)",
+    "<strong>Silver Tipped Arrows</strong> (d4 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>Crossbow</strong> (Damage: 1d6, Qualities: Missile (5’-80’ / 81’-160’ / 161’-240’), Slow, Two-handed, 3 slots)",
+    "<strong>Case of Bolts</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Short Bow</strong> (Damage: 1d6, Qualities: Missile (5’-50’ / 51’-100’ / 101’-150’), Two-handed, 2 slots)",
+    "<strong>Quiver of Arrows</strong> (d20 usage die)",
+    "<strong>Silver Tipped Arrows</strong> (d4 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ]
+];
