+++ src/characterData/equipmentKits/clericEquipment.ts
@@ -1,0 +1,115 @@
+import { getDivineScrollSpell } from "../../utils/getRandomSpellScroll";
+
+/**
+ * Cleric equipment kits cannot be higher than 9 slots.
+ * 9 + starting gold = 10
+ */
+export const clericEquipment = [
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>Shield</strong>",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Warhammer</strong> (Damage: 1d6, Qualities: Blunt, Melee, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Warhammer</strong> (Damage: 1d6, Qualities: Blunt, Melee, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Warhammer</strong> (Damage: 1d6, Qualities: Blunt, Melee, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Silver Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Silver Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Silver Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Warhammer</strong> (Damage: 1d6, Qualities: Blunt, Melee, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Warhammer</strong> (Damage: 1d6, Qualities: Blunt, Melee, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Silver Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Vial Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)"
+  ]
+];
