+++ src/characterData/equipmentKits/drowEquipment.ts
@@ -1,0 +1,112 @@
+import { getDivineScrollSpell } from "../../utils/getRandomSpellScroll";
+
+/**
+ * Drow equipment kits cannot be higher than 9 slots.
+ * 9 + starting gold = 10
+ */
+export const drowEquipment = [
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Short Sword</strong> (Damage: 1d6, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Short Sword</strong> (Damage: 1d6, Qualities: Melee, 2 slots)",
+    "<strong>Short Sword</strong> (Damage: 1d6, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Two-Handed Sword</strong> (Damage: 1d10, Qualities: Melee, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Two-Handed Sword</strong> (Damage: 1d10, Qualities: Melee, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Silver Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Crossbow</strong> (Damage: 1d6, Qualities: Missile (5’-80’ / 81’-160’ / 161’-240’), Slow, Two-handed, 3 slots)",
+    "<strong>Case of Bolts</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Silver Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Crossbow</strong> (Damage: 1d6, Qualities: Missile (5’-80’ / 81’-160’ / 161’-240’), Slow, Two-handed, 3 slots)",
+    "<strong>Case of Bolts</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Silver Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Wooden Holy Symbol</strong>",
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDivineScrollSpell()}`
+  ],
+  [
+    "<strong>Silver Holy Symbol</strong>",
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Mace</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ]
+];
