+++ src/characterData/equipmentKits/druidEquipment.ts
@@ -1,0 +1,125 @@
+import { getDruidScrollSpell } from "../../utils/getRandomSpellScroll";
+/**
+ * Druid equipment kits cannot be higher than 9 slots.
+ * They also have a holy symbol.
+ * 9 + starting gold = 10
+ */
+export const druidEquipment = [
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Shield</strong> (wooden)",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Shield</strong> (wooden)",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Spear</strong> (Damage: 1d6, Qualities: Brace, Melee, Missile (5’-20’ / 21’-40’ / 41’-60’), 2 slots)",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Shield</strong> (wooden)",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDruidScrollSpell()}`
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDruidScrollSpell()}`
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDruidScrollSpell()}`
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Cudgel</strong> (Damage: 1d4, Qualities: Blunt, Melee)",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDruidScrollSpell()}`
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Shield</strong> (wooden)",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDruidScrollSpell()}`
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDruidScrollSpell()}`
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Spear</strong> (Damage: 1d6, Qualities: Brace, Melee, Missile (5’-20’ / 21’-40’ / 41’-60’), 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDruidScrollSpell()}`
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getDruidScrollSpell()}`
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Shield</strong> (wooden)",
+    "<strong>Spear</strong> (Damage: 1d6, Qualities: Brace, Melee, Missile (5’-20’ / 21’-40’ / 41’-60’), 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Shield</strong> (wooden)",
+    "<strong>Spear</strong> (Damage: 1d6, Qualities: Brace, Melee, Missile (5’-20’ / 21’-40’ / 41’-60’), 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Spear</strong> (Damage: 1d6, Qualities: Brace, Melee, Missile (5’-20’ / 21’-40’ / 41’-60’), 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Holy Symbol</strong>",
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ]
+];
