+++ src/characterData/equipmentKits/gnomeEquipment.ts
@@ -1,0 +1,127 @@
+import { getIllusionistScrollSpell } from "../../utils/getRandomSpellScroll";
+
+/**
+ * Gnome equipment kits cannot be higher than 9 slots.
+ * 9 + starting gold = 10
+ */
+export const gnomeEquipment = [
+  [
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Short Bow</strong> (Damage: 1d6, Qualities: Missile (5’-50’ / 51’-100’ / 101’-150’), Two-handed, 2 slots)",
+    "<strong>Quiver of Arrows</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Shield</strong>",
+    "<strong>Short Sword</strong> (Damage: 1d6, Qualities: Melee, 2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Vial of Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>2 Vials Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Vial of Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Short Bow</strong> (Damage: 1d6, Qualities: Missile (5’-50’ / 51’-100’ / 101’-150’), Two-handed, 2 slots)",
+    "<strong>Quiver of Arrows</strong> (d20 usage die)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>2 Vials Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Shield</strong>",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Leather Armor</strong> (2 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Sling</strong> (Damage: 1d4, Qualities: Blunt, Missile (5’-40’ / 41’-80’ / 81’-160’))",
+    "<strong>Pouch With Sling Bullets</strong> (d20 usage die)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ]
+];
