+++ src/characterData/equipmentKits/illusionistEquipment.ts
@@ -1,0 +1,117 @@
+import { getIllusionistScrollSpell } from "../../utils/getRandomSpellScroll";
+
+/**
+ * Illusionist equipment kits cannot be higher than 9 slots.
+ * 9 + starting gold = 10
+ */
+export const illusionistEquipment = [
+  [
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Quartz</strong> (10gp)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Quartz</strong> (10gp)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Vial of Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Silver Mirror</strong>",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>2 Vials Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Vial of Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Quartz</strong> (10gp)",
+    "<strong>2 Vials Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Quartz</strong> (10gp)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    `${getIllusionistScrollSpell()}`,
+    "<strong>Quartz</strong> (10gp)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ]
+];
