+++ src/characterData/equipmentKits/knightEquipment.ts
@@ -1,0 +1,96 @@
+/**
+ * Knight equipment kits cannot be higher than 9 slots.
+ * They do not use leather armor and do not use missile
+ * weapons.
+ * 9 + starting gold = 10
+ */
+export const knightEquipment = [
+  [
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Polearm</strong> (Damage: 1d10, Qualities: Brace, Melee, Slow, Two-handed, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Flail</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Morning Star</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Battle Axe</strong> (Damage: 1d8, Qualities: Melee, Slow, Two-handed, 3 slots)",
+    "<strong>Hand Axe</strong> (Damage: 1d6, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’), 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Spear</strong> (Damage: 1d6, Qualities: Brace, Melee, Missile (5’-20’ / 21’-40’ / 41’-60’), 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Chainmail Armor</strong> (2 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Spear</strong> (Damage: 1d6, Qualities: Brace, Melee, Missile (5’-20’ / 21’-40’ / 41’-60’), 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Two-Handed Sword</strong> (Damage: 1d10, Qualities: Melee, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Flail</strong> (Damage: 1d6, Qualities: Blunt, Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Battle Axe</strong> (Damage: 1d8, Qualities: Melee, Slow, Two-handed, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Two-Handed Sword</strong> (Damage: 1d10, Qualities: Melee, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Halberd</strong> (Damage: 1d10, Qualities: Brace, Melee, Slow, Two-handed, 3 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ],
+  [
+    "<strong>Plate Armor</strong> (3 slots)",
+    "<strong>Shield</strong>",
+    "<strong>Sword</strong> (Damage: 1d8, Qualities: Melee, 2 slots)",
+    "<strong>2 Travel rations</strong> (1/day)"
+  ]
+];
