+++ src/characterData/equipmentKits/magicUserEquipment.ts
@@ -1,0 +1,112 @@
+import { getMagicUserScrollSpell } from "../../utils/getRandomSpellScroll";
+
+/**
+ * Magic-User equipment kits cannot be higher than 9 slots.
+ * 9 + starting gold = 10
+ */
+export const magicUserEquipment = [
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Vial of Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Silver Mirror</strong>",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>2 Vials Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Vial of Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>2 Vials Holy Water</strong> (Damage: 1d8, Qualities: Missile (5’-10’ / 11’-30’ / 31’-50’), Splash weapon)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>2 Travel rations</strong> (1/day)",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    "<strong>Quarter-Staff</strong> (Damage: 1d4, Qualities: Blunt, Melee, Slow, Two-handed)",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ],
+  [
+    "<strong>Silver Dagger</strong> (Damage: 1d4, Qualities: Melee, Missile (5’-10’ / 11’-20’ / 21’-30’))",
+    `${getMagicUserScrollSpell()}`,
+    "<strong>2 Travel rations</strong> (1/day)",
+    "<strong>Spellbook</strong>"
+  ]
+];
